+++ .eslintrc
@@ -15,6 +15,7 @@
     "react/jsx-key": "warn", // 반복문으로 생성하는 요소에 key 강제
     "react/self-closing-comp": "warn", // 셀프 클로징 태그 가능하면 적용
     "react/jsx-curly-brace-presence": "warn", // jsx 내 불필요한 중괄호 금지
+    "react/prop-types": "off",
     "prettier/prettier": [
       "error",
       {
