+++ src/pages/jahyunSon/Main/Main.js
@@ -1,24 +1,27 @@
-import React, { useState } from 'react';
-import Comment from '../Comment/Comment';
+import React, { useState, useEffect } from 'react';
+import Post from '../Post/Post';
 import { FOOTER_INFO_LIST } from './data';
 import './Main.scss';
 
 function MainJahyun() {
-  const [comment, setComment] = useState('');
-  const [commentList, setCommentList] = useState([]);
+  const [userList, setUserList] = useState([]);
 
-  const postComment = e => {
-    setComment(e.target.value);
-  };
+  useEffect(() => {
+    fetchUserList();
+  }, []);
 
-  const pushCommentList = e => {
-    e.preventDefault();
-    setCommentList(commentList => [...commentList, comment]);
-    setComment('');
+  const fetchUserList = () => {
+    fetch('/data/sample.json', {
+      method: 'GET',
+    })
+      .then(res => res.json())
+      .then(data => {
+        setUserList(data);
+      });
   };
 
   return (
-    <body className="main">
+    <div className="main">
       <nav>
         <div className="navBarLogo">
           <img
@@ -37,129 +40,35 @@ function MainJahyun() {
           />
         </div>
         <div className="iconsOnRight">
-          <a href="#">
+          <button>
             <img
               src="/images/jahyunSon/explore.png"
               alt="explore page"
               className="icon"
             />
-          </a>
-          <a href="#">
+          </button>
+          <button>
             <img
               src="/images/jahyunSon/heart.png"
               alt="notificaions"
               className="icon"
             />
-          </a>
-          <a href="#">
+          </button>
+          <button>
             <img
               src="/images/jahyunSon/profile.png"
               alt="my profile"
               className="icon"
             />
-          </a>
+          </button>
         </div>
       </nav>
 
       <main>
         <section className="feed">
-          <article className="post">
-            <header className="postHeader">
-              <div className="userInfo">
-                <img
-                  src="/images/jahyunSon/feedProfile.jpeg"
-                  alt="profile"
-                  className="pfp"
-                />
-                <p className="username bold">hand_jahyun</p>
-              </div>
-              <img
-                src="/images/jahyunSon/threedots.png"
-                alt="view more"
-                className="icon"
-              />
-            </header>
-            <img
-              src="/images/jahyunSon/feed1.png"
-              alt="feed"
-              className="feedImg"
-            />
-            <div className="interactionBar">
-              <div className="interactionBarLeft">
-                <img
-                  src="/images/jahyunSon/redheart.png"
-                  alt="like"
-                  className="icon"
-                />
-                <img
-                  src="/images/jahyunSon/comment.png"
-                  alt="comment"
-                  className="icon"
-                />
-                <img
-                  src="/images/jahyunSon/upload.png"
-                  alt="share"
-                  className="icon"
-                />
-              </div>
-              <img
-                src="/images/jahyunSon/bookmark.png"
-                alt="bookmark"
-                className="icon"
-              />
-            </div>
-            <div className="likedListContainer">
-              <img
-                src="/images/jahyunSon/pfp1.jpeg"
-                alt="profile"
-                className="pfpSmall"
-              />
-              <p className="likedList">
-                <span className="bold">taylorswift</span>님 외{' '}
-                <span className="bold">10명</span>이 좋아합니다
-              </p>
-            </div>
-            <p className="caption">
-              <span className="username bold">hand_jahyun</span> 제주도 바다~...{' '}
-              <span className="light">더 보기</span>
-            </p>
-            <section className="commentSection">
-              <div className="commentsList">
-                <div className="commentAndBtns" id="likeAndDelete">
-                  <div className="comment">
-                    <span className="bold floatLeft">taylorswift</span> Wow
-                    beautiful~
-                  </div>
-                  <button
-                    className="deleteBtn lighter hidden"
-                    type="button"
-                    id="deleteBtn0"
-                  >
-                    삭제
-                  </button>
-                  <button className="likeBtn" type="button" id="likeBtn">
-                    <i className="fa fa-heart" id="likeBtn"></i>
-                  </button>
-                </div>
-                <Comment comment={comment} commentList={commentList} />
-              </div>
-              <form className="postComment">
-                <input
-                  type="text"
-                  className="commentInput light"
-                  onChange={e => postComment(e)}
-                  value={comment}
-                />
-                <button
-                  className="commentBtn"
-                  type="submit"
-                  onClick={pushCommentList}
-                >
-                  게시
-                </button>
-              </form>
-            </section>
-          </article>
+          {userList.map(user => {
+            return <Post key={user.id} username={user.username} />;
+          })}
         </section>
 
         <aside className="main-right">
@@ -269,7 +178,7 @@ function MainJahyun() {
           </footer>
         </aside>
       </main>
-    </body>
+    </div>
   );
 }
