+++ src/pages/jahyunSon/Main/Main.scss
@@ -85,118 +85,6 @@
 
     .feed {
       margin: 0 20px 0 50px;
-
-      .post {
-        width: 470px;
-        border: 1px solid #e6e6e6;
-        background-color: white;
-
-        .postHeader {
-          @include flexSort(row, space-between, center);
-          margin: 8px 4px 8px 5px;
-
-          .userInfo {
-            @include flexSort(row, flex-start, center);
-
-            .pfp {
-              width: 32px;
-              margin-left: 8px;
-              border-radius: 50%;
-            }
-
-            .username {
-              margin: 0 0 0 12px;
-            }
-          }
-        }
-
-        .feedImg {
-          width: 468px;
-        }
-
-        .interactionBar {
-          @include flexSort(row, space-between, center);
-          margin: 0 5px 0 10px;
-        }
-
-        .likedListContainer {
-          @include flexSort(row, flex-start, center);
-          height: 20px;
-          margin: 0 0 6px 12px;
-
-          .pfpSmall {
-            width: 15px;
-            margin-right: 4px;
-            border-radius: 50%;
-          }
-        }
-
-        .caption {
-          margin: 2px 12px;
-        }
-
-        .commentSection {
-          @include flexSort(column, center, flex-start);
-
-          .commentsList {
-            margin: 1px 0 1px 10px;
-            padding: 0;
-
-            .commentAndBtns {
-              @include flexSort(row, flex-start, flex-start);
-              //align-items: top;
-
-              .comment {
-                width: 400px;
-                margin: 3px 0;
-                float: left;
-                word-wrap: break-word;
-              }
-
-              .deleteBtn {
-                width: 20px;
-                height: 20px;
-                margin: 0px 5px;
-              }
-
-              .likeBtn {
-                width: 10px;
-                height: 10px;
-                margin: 3px 5px;
-              }
-            }
-          }
-
-          .postComment {
-            width: 100%;
-            height: 38px;
-            margin-top: 5px;
-            border-top: 1px solid #e6e6e6;
-
-            .commentInput {
-              width: 360px;
-              height: 36px;
-              padding: 0 12px;
-              border-style: none;
-              outline: none;
-              text-align: left;
-            }
-
-            .commentBtn {
-              padding: 11px 12px;
-              float: right;
-              border-style: none;
-              background-color: white;
-              color: #c8e2fb;
-
-              &:hover,
-              &:focus {
-                color: #5eacd6;
-              }
-            }
-          }
-        }
-      }
     }
   }
